+++ lib/test/cases/u_strings.py
@@ -1,0 +1,9 @@
+u"u-strings were a Python 2.0 thing that hung about for compatibility"
+
+u"""
+They're redundant now since all strings are unicode by default, but we should make sure
+we don't mess any thing up.
+
+See also:
+https://stackoverflow.com/a/2464968
+"""