+++ lib/test/snapshots/r_strings.py
@@ -1,0 +1,9 @@
+r'abcdef'
+
+r'abc\"def\"'
+
+r'abc\'def\'ghi'
+
+rf'abc{"def"}ghi'
+
+rf'abc{r"d\'ef"}ghi'