+++ lib/CHANGELOG.md
@@ -7,6 +7,7 @@
 - `include` config option for overriding `ignore`
 - Support for populating the `ignore` config from `.gitignore` files
 - Sensible default value for `ignore` config option
+- Support for r-strings, rf-strings and even u-strings
 
 ### Fixed